refactor: remove DOMAIN_HEURISTICS dead code, rename heuristics.ts → constants.ts
- heuristics.ts renamed to constants.ts: DOMAIN_HEURISTICS removed
  (was only used by categorizer.ts which was deleted in a previous commit)
- All imports updated from heuristics.js to constants.js
- scripts/sync-index.mjs removed: redundant since sync-skills.yml
  already downloads skills_index.json via curl during CI
