+++ src/skill-pointer/pointer-generator.ts
@@ -1,18 +1,9 @@
 import fs from "fs";
 import path from "path";
-import { POINTER_SUFFIX, SKILL_FILENAME } from "../constants/heuristics.js";
+import { POINTER_SUFFIX, SKILL_FILENAME } from "../constants/constants.js";
 import { ensureDir, listSubdirectories } from "../utils/fs.utils.js";
 import type { SkillIndexEntry } from "./vault-installer.js";
 
-/**
- * Builds the SKILL.md content for a category pointer.
- *
- * The pointer includes:
- * - Instructions for the AI to browse the vault on demand.
- * - A full list of skill names + descriptions so that keyword-based
- *   tools like get_available_skills can match queries like
- *   "laravel", "wordpress", "react", etc. without loading every SKILL.md.
- */
 function buildPointerContent(
   category: string,
   skills: SkillIndexEntry[],
@@ -60,9 +51,9 @@ ${skillList}
  * Scans every category directory in the vault and writes
  * a lightweight pointer SKILL.md into the active skills directory.
  *
- * If skills_index.json entries are available for a category, the pointer
- * will include the full skill list for searchability. Otherwise it falls
- * back to counting files in the vault directory.
+ * Each pointer includes the full list of skill names + descriptions
+ * so keyword searches (e.g. "laravel", "wordpress") resolve correctly
+ * via get_available_skills without loading every SKILL.md.
  */
 export function generatePointers(
   activeSkillsDir: string,
@@ -71,7 +62,6 @@ export function generatePointers(
 ): void {
   const categoryDirs = listSubdirectories(vaultDir);
 
-  // Group index entries by category for O(1) access per category
   const byCategory = new Map<string, SkillIndexEntry[]>();
   for (const entry of index) {
     const cat = entry.category ?? "_uncategorized";
@@ -83,11 +73,10 @@ export function generatePointers(
     const categoryVaultPath = path.join(vaultDir, categoryName);
     const skills = byCategory.get(categoryName) ?? [];
 
-    // Skip empty vault category dirs that have no indexed skills either
     if (skills.length === 0) {
-      const subDirs = fs.readdirSync(categoryVaultPath).filter((e) => {
-        return fs.statSync(path.join(categoryVaultPath, e)).isDirectory();
-      });
+      const subDirs = fs.readdirSync(categoryVaultPath).filter((e) =>
+        fs.statSync(path.join(categoryVaultPath, e)).isDirectory()
+      );
       if (subDirs.length === 0) continue;
     }
 
@@ -97,8 +86,10 @@ export function generatePointers(
     );
 
     ensureDir(pointerDir);
-
-    const content = buildPointerContent(categoryName, skills, categoryVaultPath);
-    fs.writeFileSync(path.join(pointerDir, SKILL_FILENAME), content, "utf-8");
+    fs.writeFileSync(
+      path.join(pointerDir, SKILL_FILENAME),
+      buildPointerContent(categoryName, skills, categoryVaultPath),
+      "utf-8"
+    );
   }
 }
