+++ src/skill-pointer/vault-installer.ts
@@ -1,7 +1,7 @@
 import fs from "fs";
 import path from "path";
-import { SKILL_FILENAME } from "../constants/heuristics.js";
 import { ensureDir } from "../utils/fs.utils.js";
+import { UNCATEGORIZED_CATEGORY } from "../constants/constants.js";
 
 export interface SkillIndexEntry {
   id: string;
@@ -13,7 +13,7 @@ export interface SkillIndexEntry {
 /**
  * Loads the pre-built skills_index.json that ships alongside bundled-skills.
  * This index is generated by the upstream Antigravity repo (generate_index.py)
- * and downloaded during the sync-skills build step.
+ * and downloaded during the CI sync-skills workflow step.
  *
  * Falls back to an empty array if the file is missing so the plugin degrades
  * gracefully without crashing OpenCode.
@@ -47,16 +47,21 @@ export function installSkillsToVault(
 ): void {
   if (!fs.existsSync(bundledSkillsPath)) return;
 
-  // Build O(1) lookup: skill folder name → category
-  const categoryMap = new Map(index.map((e) => [e.id, e.category ?? "_uncategorized"]));
+  const categoryMap = new Map(
+    index.map((e) => [e.id, e.category ?? UNCATEGORIZED_CATEGORY])
+  );
 
   for (const entry of fs.readdirSync(bundledSkillsPath)) {
-    // Skip hidden files, the index file itself, and non-directories
-    if (entry.startsWith(".") || entry === "skills_index.json" || entry === "README.md") continue;
+    if (
+      entry.startsWith(".") ||
+      entry === "skills_index.json" ||
+      entry === "README.md"
+    ) continue;
+
     const srcPath = path.join(bundledSkillsPath, entry);
     if (!fs.statSync(srcPath).isDirectory()) continue;
 
-    const category = categoryMap.get(entry) ?? "_uncategorized";
+    const category = categoryMap.get(entry) ?? UNCATEGORIZED_CATEGORY;
     const destPath = path.join(vaultDir, category, entry);
 
     ensureDir(path.join(vaultDir, category));
